Move `OneofDeclaration` to `io.spine.code.proto`
