+++ base/src/main/java/io/spine/validate/MessageValue.java
@@ -27,8 +27,8 @@ import com.google.protobuf.Descriptors.OneofDescriptor;
 import com.google.protobuf.Message;
 import io.spine.code.proto.FieldContext;
 import io.spine.code.proto.FieldDeclaration;
+import io.spine.code.proto.OneofDeclaration;
 import io.spine.type.MessageType;
-import io.spine.type.OneofDeclaration;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.util.Optional;
