+++ base/src/main/java/io/spine/validate/option/IsRequired.java
@@ -22,7 +22,7 @@ package io.spine.validate.option;
 
 import com.google.errorprone.annotations.Immutable;
 import com.google.protobuf.Descriptors.OneofDescriptor;
-import io.spine.type.OneofDeclaration;
+import io.spine.code.proto.OneofDeclaration;
 import io.spine.validate.Constraint;
 
 import java.util.Optional;
