+++ base/src/main/java/io/spine/validate/option/IsRequiredConstraint.java
@@ -23,8 +23,8 @@ package io.spine.validate.option;
 import com.google.errorprone.annotations.Immutable;
 import io.spine.code.proto.FieldContext;
 import io.spine.code.proto.FieldName;
+import io.spine.code.proto.OneofDeclaration;
 import io.spine.type.MessageType;
-import io.spine.type.OneofDeclaration;
 import io.spine.validate.Constraint;
 import io.spine.validate.ConstraintTranslator;
